Revised the Wiki pages in favor of better instructions for the Kobra 2 pro (#200)
* Rewrote the install instructions to spotlight the process of installing Rinkhals on the Kobra 2 pro.

Split the instructions for each series to make it easier to read.
Added the correct folder name to use with the Kobra 2 pro.
Added the correct directions to install Rinkhals on the Kobra 2 pro.

* Updated the quick start page to reflect the current install processes.

Split the instructions per series.
Added a link to the install page to avoid repeated speach.

* Created simple and basic instructions for app installation.

Wrote simple instructions for installing apps using the swu files.
Added a separate line about using apps not from the official repo.
Added a note telling users to confirm the app works with their specific printer.

* Updated the readme to be consistent with the new quick guide page.